move functions to their respective files
